requirements.txt
Werkzeug==2.0.1
numpy==1.23.4 
pandas==1.5.1
joblib==1.2.0
scikit-learn==1.0.2
flask==2.0.1
WTForms==3.0.1
Flask-WTF==1.0.1
python-dotenv==0.20.0
flask-sqlalchemy==2.5.1
flask-login==0.6.2
sqlalchemy==1.4.40
gunicorn==20.1.0
pydantic
alembic
Flask-Migrate
psycopg2-binary